chore: move metric toolset to metrics package
No functional change.
