+++ pkg/metrics/prompt.go
@@ -1,4 +1,4 @@
-package tools
+package metrics
 
 const (
 	ServerPrompt = `You are an expert Kubernetes and OpenShift observability assistant with direct access to Prometheus metrics and Alertmanager alerts through this MCP server.
