+++ pkg/metrics/summary.go
@@ -1,4 +1,4 @@
-package tools
+package metrics
 
 import (
 	"math"
