+++ pkg/metrics/summary_test.go
@@ -1,4 +1,4 @@
-package tools
+package metrics
 
 import (
 	"math"
