+++ pkg/metrics/tooldef.go
@@ -1,4 +1,4 @@
-package tools
+package metrics
 
 import (
 	"encoding/json"
@@ -7,18 +7,8 @@ import (
 
 	"github.com/containers/kubernetes-mcp-server/pkg/api"
 	"github.com/google/jsonschema-go/jsonschema"
-
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 )
-
-// MustSchema returns the JSON Schema for type T, panicking on error.
-func MustSchema[T any]() *jsonschema.Schema {
-	s, err := jsonschema.For[T](nil)
-	if err != nil {
-		panic(err)
-	}
-	return s
-}
 
 // ToolDefInterface defines the common interface for all tool definitions
 type ToolDefInterface interface {
