+++ pkg/metrics/tooldef_test.go
@@ -1,4 +1,4 @@
-package tools
+package metrics
 
 import (
 	"testing"
